+++ src/services/recruiting/index.ts
@@ -2,3 +2,4 @@
 export * from "./recruitingService";
 export * from "./recruitInvitationService";
 export * from "./brandingSettingsService";
+export * from "./authUserService";